requirements.txt
Django==3.2.19
djangorestframework==3.14.0
transformers==4.30.2
